Switch to PostgreSQL, add initial EF migration
- Migrated from SQL Server to PostgreSQL (Npgsql) as the database provider; updated connection strings and removed SQL Server dependencies.
- Updated EF Core and related packages to v10.0.0.
- Added initial EF Core migration and model snapshot for PostgreSQL schema.
- Changed ShipmentItemEntity.Qty from decimal to int.
- Made EmployeeEntity.premiumPercent nullable and updated validation.
- Registered ICurrentUserService and InitialDataSeeder in DI; added default CurrentUserService and InitialDataSeeder implementations.
- Uncommented MediatR registration in Program.cs.
- Downgraded AutoMapper.Extensions.Microsoft.DependencyInjection to 12.0.0.
- Minor code and namespace cleanups.

+++ src/MyFactory.Domain/Entities/Orders/OrdersEntities.cs
@@ -224,7 +224,7 @@ public class ShipmentItemEntity : AuditableEntity
     public Guid SalesOrderItemId { get; private set; }
     public Guid WarehouseId { get; private set; }
     public Guid ProductId { get; private set; }
-    public decimal Qty { get; private set; }
+    public int Qty { get; private set; }
     public decimal UnitPrice { get; private set; }
 
     // Navigation properties
